+++ packages/react/src/components/ContextMenu/index.tsx
@@ -385,6 +385,89 @@ const ContextMenu: React.FC = () => {
               </Menu>
             ));
       }
+      if (name === "insert-column-right") {
+        return selection?.row_select
+          ? null
+          : ["left"].map((dir) => (
+              <Menu
+                key={`add-col-${dir}`}
+                onClick={() => {
+                  const position =
+                    context.luckysheet_select_save?.[0]?.column?.[0];
+                  if (position == null) return;
+                  // const countStr = (e.target as HTMLDivElement).querySelector(
+                  //   "input"
+                  // )?.value;
+                  // if (countStr == null) return;
+                  // const count = parseInt(countStr, 10);
+                  const count = 1;
+                  if (count < 1) return;
+                  const direction = "rightbottom";
+                  const insertRowColOp: SetContextOptions["insertRowColOp"] = {
+                    type: "column",
+                    index: position,
+                    count,
+                    direction,
+                    id: context.currentSheetId,
+                  };
+                  setContext(
+                    (draftCtx) => {
+                      try {
+                        insertRowCol(draftCtx, insertRowColOp);
+                        draftCtx.contextMenu = {};
+                      } catch (err: any) {
+                        if (err.message === "maxExceeded")
+                          showAlert(rightclick.columnOverLimit, "ok");
+                        else if (err.message === "readOnly")
+                          showAlert(
+                            rightclick.cannotInsertOnColumnReadOnly,
+                            "ok"
+                          );
+                        draftCtx.contextMenu = {};
+                      }
+                    },
+                    {
+                      insertRowColOp,
+                    }
+                  );
+                }}
+              >
+                <div className="context-item">
+                  <LucideIcon name="Plus" />
+                  <div>
+                    Insert column to the right
+                    {/* {_.startsWith(context.lang ?? "", "zh") && (
+                      <>
+                        {rightclick.to}
+                        <span className={`luckysheet-cols-rows-shift-${dir}`}>
+                          {(rightclick as any)[dir]}
+                        </span>
+                      </>
+                    )}
+                    {`${rightclick.insert}  `} */}
+                    {/* <input
+                  onClick={(e) => e.stopPropagation()}
+                  onKeyDown={(e) => e.stopPropagation()}
+                  tabIndex={0}
+                  type="text"
+                  className="luckysheet-mousedown-cancel"
+                  placeholder={rightclick.number}
+                  defaultValue="1"
+                /> */}
+                    {/* 1
+                    <span className="luckysheet-cols-rows-shift-word luckysheet-mousedown-cancel">
+                      {` ${rightclick.column}  `}
+                    </span>
+                    {!_.startsWith(context.lang ?? "", "zh") && (
+                      <span className={`luckysheet-cols-rows-shift-${dir}`}>
+                        {(rightclick as any)[dir]}
+                      </span>
+                    )} */}
+                  </div>
+                </div>
+              </Menu>
+            ));
+      }
       if (name === "insert-row") {
         return selection?.column_select
           ? null
@@ -429,6 +512,81 @@ const ContextMenu: React.FC = () => {
                   <LucideIcon name="Plus" />
                   <div>
                     Insert row below
+                    {/* {_.startsWith(context.lang ?? "", "zh") && (
+                      <>
+                        {rightclick.to}
+                        <span className={`luckysheet-cols-rows-shift-${dir}`}>
+                          {(rightclick as any)[dir]}
+                        </span>
+                      </>
+                    )}
+                    {`${rightclick.insert}  `}1 */}
+                    {/* <input
+                  onClick={(e) => e.stopPropagation()}
+                  onKeyDown={(e) => e.stopPropagation()}
+                  tabIndex={0}
+                  type="text"
+                  className="luckysheet-mousedown-cancel"
+                  placeholder={rightclick.number}
+                  defaultValue="1"
+                /> */}
+                    {/* <span className="luckysheet-cols-rows-shift-word luckysheet-mousedown-cancel">
+                      {` ${rightclick.row}  `}
+                    </span>
+                    {!_.startsWith(context.lang ?? "", "zh") && (
+                      <span className={`luckysheet-cols-rows-shift-${dir}`}>
+                        {(rightclick as any)[dir]}
+                      </span>
+                    )} */}
+                  </div>
+                </div>
+              </Menu>
+            ));
+      }
+      if (name === "insert-row-above") {
+        return selection?.column_select
+          ? null
+          : ["bottom"].map((dir) => (
+              <Menu
+                key={`add-row-${dir}`}
+                onClick={() => {
+                  const position =
+                    context.luckysheet_select_save?.[0]?.row?.[0];
+                  if (position == null) return;
+                  // const countStr = container.querySelector("input")?.value;
+                  // if (countStr == null) return;
+                  const count = 1;
+                  // const count = parseInt(countStr, 10);
+                  if (count < 1) return;
+                  const direction = "lefttop";
+                  const insertRowColOp: SetContextOptions["insertRowColOp"] = {
+                    type: "row",
+                    index: position,
+                    count,
+                    direction,
+                    id: context.currentSheetId,
+                  };
+                  setContext(
+                    (draftCtx) => {
+                      try {
+                        insertRowCol(draftCtx, insertRowColOp);
+                        draftCtx.contextMenu = {};
+                      } catch (err: any) {
+                        if (err.message === "maxExceeded")
+                          showAlert(rightclick.rowOverLimit, "ok");
+                        else if (err.message === "readOnly")
+                          showAlert(rightclick.cannotInsertOnRowReadOnly, "ok");
+                        draftCtx.contextMenu = {};
+                      }
+                    },
+                    { insertRowColOp }
+                  );
+                }}
+              >
+                <div className="context-item">
+                  <LucideIcon name="Plus" />
+                  <div>
+                    Insert row above
                     {/* {_.startsWith(context.lang ?? "", "zh") && (
                       <>
                         {rightclick.to}