+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @fileverse-dev/fortune-core
 
+## 1.1.77
+
+### Patch Changes
+
+- N/A
+
+## 1.1.76
+
+### Patch Changes
+
+- minor changes for stats
+
+## 1.1.75
+
+### Patch Changes
+
+- footer UI changes
+
 ## 1.1.74
 
 ### Patch Changes